+++ lib/src/Payment.dart
@@ -2,6 +2,7 @@ import '../cielo_ecommerce.dart';
 
 part 'Payment.g.dart';
 
+///Payment
 class Payment {
   int serviceTaxAmount;
   int installments;
@@ -32,37 +33,59 @@ class Payment {
   int capturedAmount;
   String voidedDate;
   int voidedAmount;
+  String address;
+  String boletoNumber;
+  String assignor;
+  String demonstrative;
+  String expirationDate;
+  String identification;
+  String instructions;
+  String url;
+  String barCodeNumber;
+  String digitableLine;
+  int bank;
 
   Payment(
       {this.serviceTaxAmount,
-      this.installments,
-      this.interest,
-      this.capture,
-      this.authenticate,
-      this.recurrent,
-      this.creditCard,
-      this.tid,
-      this.proofOfSale,
-      this.authorizationCode,
-      this.softDescriptor,
-      this.provider,
-      this.fraudAnalysis,
-      this.isQrCode,
-      this.amount,
-      this.receivedDate,
-      this.status,
-      this.isSplitted,
-      this.returnMessage,
-      this.returnCode,
-      this.paymentId,
-      this.type,
-      this.currency,
-      this.country,
-      this.links,
-      this.capturedDate,
-      this.capturedAmount,
-      this.voidedDate,
-      this.voidedAmount});
+        this.installments,
+        this.interest,
+        this.capture,
+        this.authenticate,
+        this.recurrent,
+        this.creditCard,
+        this.tid,
+        this.proofOfSale,
+        this.authorizationCode,
+        this.softDescriptor,
+        this.provider,
+        this.fraudAnalysis,
+        this.isQrCode,
+        this.amount,
+        this.receivedDate,
+        this.status,
+        this.isSplitted,
+        this.returnMessage,
+        this.returnCode,
+        this.paymentId,
+        this.type,
+        this.currency,
+        this.country,
+        this.links,
+        this.capturedDate,
+        this.capturedAmount,
+        this.voidedDate,
+        this.voidedAmount,
+        this.address,
+        this.boletoNumber,
+        this.assignor,
+        this.demonstrative,
+        this.expirationDate,
+        this.identification,
+        this.instructions,
+        this.url,
+        this.digitableLine,
+        this.barCodeNumber,
+        this.bank});
 
   factory Payment.fromJson(Map<String, dynamic> json) =>
       _$PaymentFromJson(json);